requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
attrs==21.4.0
    # via
    #   cattrs
    #   jsii
avro==1.11.0
    # via -r requirements.in
aws-cdk-aws-msk-alpha==2.22.0a0
    # via -r requirements.in
aws-cdk-lib==2.22.0
    # via
    #   -r requirements.in
    #   aws-cdk-aws-msk-alpha
black==22.3.0
    # via -r requirements.in
boto3==1.22.3
    # via -r requirements.in
boto3-stubs==1.22.3
    # via -r requirements.in
botocore==1.25.3
    # via
    #   boto3
    #   s3transfer
botocore-stubs==1.25.3
    # via boto3-stubs
cattrs==22.1.0
    # via jsii
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   black
    #   pip-tools
constructs==10.0.127
    # via
    #   aws-cdk-aws-msk-alpha
    #   aws-cdk-lib
exceptiongroup==1.0.0rc4
    # via cattrs
idna==3.3
    # via requests
isort==5.10.1
    # via -r requirements.in
jmespath==1.0.0
    # via
    #   boto3
    #   botocore
jsii==1.57.0
    # via
    #   aws-cdk-aws-msk-alpha
    #   aws-cdk-lib
    #   constructs
kafka-python==2.0.2
    # via -r requirements.in
mypy==0.950
    # via -r requirements.in
mypy-boto3-glue==1.22.1
    # via -r requirements.in
mypy-boto3-kafka==1.22.0.post1
    # via -r requirements.in
mypy-boto3-ssm==1.22.0.post1
    # via -r requirements.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pip-tools==6.6.0
    # via -r requirements.in
platformdirs==2.5.2
    # via black
publication==0.0.3
    # via
    #   aws-cdk-aws-msk-alpha
    #   aws-cdk-lib
    #   constructs
python-dateutil==2.8.2
    # via
    #   botocore
    #   jsii
pyyaml==6.0
    # via -r requirements.in
requests==2.27.1
    # via v20
s3transfer==0.5.2
    # via boto3
six==1.16.0
    # via python-dateutil
tomli==2.0.1
    # via
    #   black
    #   mypy
    #   pep517
typing-extensions==4.2.0
    # via
    #   boto3-stubs
    #   botocore-stubs
    #   jsii
    #   mypy
    #   mypy-boto3-glue
    #   mypy-boto3-kafka
    #   mypy-boto3-ssm
ujson==5.2.0
    # via v20
urllib3==1.26.9
    # via
    #   botocore
    #   requests
v20==3.0.25.0
    # via -r requirements.in
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
